requirements.txt
alpaca-trade-api
yfinance
flask
pandas
numpy
scikit-learn
python-dotenv
schedule
websocket-client
